Corrections in locations
